+++ pymongo/common.py
@@ -508,7 +508,9 @@ KW_VALIDATORS = {
     'document_class': validate_document_class,
     'read_preference': validate_read_preference,
     'event_listeners': _validate_event_listeners,
-    'tzinfo': validate_tzinfo
+    'tzinfo': validate_tzinfo,
+    'username': validate_string,
+    'password': validate_string,
 }
 
 URI_VALIDATORS.update(TIMEOUT_VALIDATORS)